+ gcc 14 doesn't want me to be able to set niceness on linux.  macro'ed out the nice() commands in util.c and driver.c when using gcc14 until I can find away around it.
+ defined several different CUDA defines instead of just one (i.e., HAVE_CUDA becomes HAVE_CUDA_LA, HAVE_CUDA_POLY, and HAVE_CUDA_BATCH) to better support which tools have cuda enabled.
+ improvements to some of the low-level ptx intrinsics for doing fixed-width Montgomery multiplications
+ support wider variety of input parameters to gpu_cofactorization (b1/b2, num curves, etc)
+ defined a new optional and experimental build macro for lasieve5_64 that attempts to optimize the operation of lambda when sieving within the factor base.  to try it out, define (uncomment) the FIXED_MAXLOG macro on line 1274 of gnfs-lasieve4e.c
+ gcc 14 is much more picky about some things.  added headers and POSIX defines here and there to satisfy it.  Have not checked for backwards compatibility issues yet with older versions of gcc.

+++ factor/qs/tdiv.c
@@ -273,13 +273,13 @@ void trial_divide_Q_siqs(uint32_t report_num,  uint8_t parity,
 		(mpz_cmp_ui(dconf->Qvals[report_num], sconf->large_prime_max) < 0))
 	{
         // save this slp (single large prime)
-		uint32_t large_prime[4];
+		uint64_t large_prime[4];
 		
-		large_prime[0] = (uint32_t)mpz_get_ui(dconf->Qvals[report_num]); //Q->val[0];
+		large_prime[0] = mpz_get_ui(dconf->Qvals[report_num]); //Q->val[0];
 		//if (dconf->num_alp == 1)
         //    large_prime[1] = dconf->curr_poly->qlisort[dconf->curr_poly->s - 1];
         //else
-            large_prime[1] = 1;
+        large_prime[1] = 1;
 		large_prime[2] = 1;
         large_prime[3] = 1;
 
@@ -486,14 +486,14 @@ void trial_divide_Q_siqs(uint32_t report_num,  uint8_t parity,
 			
 			if ((f64 > 1) && (f64 != q64) && ((q64 % f64) == 0))
 			{
-				uint32_t large_prime[4];
+				uint64_t large_prime[4];
 
-                if ((f64 < (uint64_t)sconf->large_prime_max) && 
-                    ((q64 / f64) < (uint64_t)sconf->large_prime_max))
+                if ((f64 < sconf->large_prime_max) && 
+                    ((q64 / f64) < sconf->large_prime_max))
 				{
 					//add this one
-                    large_prime[0] = (uint32_t)f64;
-                    large_prime[1] = (uint32_t)(q64 / f64);
+                    large_prime[0] = f64;
+                    large_prime[1] = (q64 / f64);
                     if (dconf->num_alp == 1)
                     {
                         large_prime[2] = dconf->curr_poly->qlisort[dconf->curr_poly->s - 1];
@@ -943,7 +943,7 @@ void trial_divide_Q_siqs(uint32_t report_num,  uint8_t parity,
 	return;
 }
 
-void buffer_relation(uint32_t offset, uint32_t *large_prime, uint32_t num_factors,
+void buffer_relation(uint32_t offset, uint64_t *large_prime, uint32_t num_factors,
     uint32_t *fb_offsets, uint32_t apoly_id, uint32_t bpoly_id, uint32_t parity,
     dynamic_conf_t *conf, uint32_t *polya_factors,
     uint32_t num_polya_factors, uint64_t unfactored_residue)
